requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=h7f98852_5
brotli-bin=1.0.9=h7f98852_5
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py39hf3d152e_1
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.28=pypi_0
dbus=1.13.6=he372182_0
einops=0.4.0=pypi_0
expat=2.2.10=h9c3ff4c_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
grad-cam=1.3.7=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=h28cd5cc_2
h5py=3.6.0=pypi_0
icu=58.2=hf484d3e_1000
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
kiwisolver=1.3.2=py39h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libbrotlicommon=1.0.9=h7f98852_5
libbrotlidec=1.0.9=h7f98852_5
libbrotlienc=1.0.9=h7f98852_5
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.13=h7f98852_1003
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.1=py39hf3d152e_0
matplotlib-base=3.5.1=py39ha18d171_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numexpr=2.8.1=py39h6abb31d_0
numpy=1.22.2=pypi_0
numpy-base=1.21.2=py39h79a1101_0
olefile=0.46=pyhd3eb1b0_0
opencv-python=4.5.5.62=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1m=h7f8727e_0
packaging=20.4=py_0
pandas=1.4.1=py39h295c915_0
pcre=8.45=h9c3ff4c_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
ptflops=0.6.9=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
pyparsing=2.4.7=py_0
pyqt=5.9.2=py39h2531618_6
python=3.9.7=h12debd9_1
python-dateutil=2.8.1=py_0
python_abi=3.9=2_cp39
pytorch=1.10.2=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2020.1=py_0
pyyaml=6.0=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
setuptools=58.0.4=py39h06a4308_0
sip=4.19.13=py39h295c915_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=hc218d9a_0
threadpoolctl=3.1.0=pypi_0
timm=0.5.4=pypi_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.10.2=py39_cu113
torchvision=0.11.3=py39_cu113
tornado=6.1=py39h3811e60_1
tqdm=4.62.3=pypi_0
ttach=0.0.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
wheel=0.37.1=pyhd3eb1b0_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
